requirements.txt
fastapi == 0.103.2
uvicorn[standard] == 0.23.2
strawberry-graphql[fastapi] == 0.209.2
pytest == 7.4.2
pytest-asyncio == 0.21.1
pylint == 2.17.5
black == 23.9.1
